+++ fotoskazka/tests/Feature/Filament/MediaUploadTest.php
@@ -19,6 +19,8 @@ class MediaUploadTest extends TestCase
     {
         parent::setUp();
 
+        $this->signInAsAdmin();
+
         Storage::fake('public');
         Storage::fake('thumbnails');
         Storage::fake('image_cache');